+++ quest/index.html
@@ -10,8 +10,8 @@
 </head>
 <main> 
     <form id="question-choices-container">
+        <h1 id="title"></h1>
         <img id="img">
-        <span id="title"></span>
         <p id="questions"></p>
         <ul id="the-choices"></ul>
         <button>Submit</button>
