+++ quest/quest.js
@@ -1,5 +1,6 @@
 import quests from '../data/questData.js';
 import { findById } from './findById.js';
+import { createChoice } from './createChoice.js';
 // import loadProfile from '../utils/loadProfile.js';
 // // import { getUser } from '../utils/loadProfile.js';
 // loadProfile();
@@ -11,30 +12,16 @@ if (!quest) {
     window.location = '../map';
 }
 
-    //grab question-choices from DOM
-const form = document.getElementById('question-choices-container');
-const img = document.getElementById('img');
 const title = document.getElementById('title');
+title.textContent = quest.title;
+const img = document.getElementById('img');
 const question = document.getElementById('questions');
+question.textContent = quest.description;
 const userChoices = document.getElementById('the-choices');
 img.src = '../assets/quest/' + quest.image;
-question.textContent = quest.description;
 for (let i = 0; i < quest.choices.length; i++){
     const choice = quest.choices[i];
     const choiceDom = createChoice(choice);
     userChoices.appendChild(choiceDom);
 }
-function createChoice(choice) {
-    const label = document.createElement('labal');
-    label.classList.add('choice');
-    const radio = document.createElement('input');
-    label.appendChild(radio);
-    radio.type = 'radio';
-    radio.name = 'choice';
-    radio.required = 'true';
-    radio.id = choice.id;
-    const description = document.createElement('span');
-    description.textContent = choice.description;
-    label.appendChild(description);
-    return label;
-}
+
